requirements.txt
channels
Django
dj-database-url
daphne
django-cors-headers
django-rest-swagger
djangorestframework
djangorestframework-simplejwt
drf-yasg
Faker
gunicorn
psycopg2
psycopg2-binary
python-decouple
django-extensions
django-filter
Pillow
django-storages
boto3